+++ test/functional-tests/node-transformers/rename-properties-transformers/rename-properties-transformer/fixtures/duplicated-generated-names-1.js
@@ -1,0 +1,4 @@
+var object = {
+    b: 'field',
+    bar: 'value'
+};