+++ CHANGELOG.md
@@ -5,6 +5,7 @@ v2.16.0
 * `stringArrayWrappersType: 'function'` now generates different indexes between each wrapper inside the same lexical scope
 * `stringArrayWrappersType: 'function'` now generates different parameters order between each wrapper inside the same lexical scope
 * `stringArrayWrappersType: 'function'` now appends `FunctionDeclaration` functions instead of `FunctionExpression` functions. This allows to append these wrappers at random positions inside each scope
+* `renameProperties` option now won't generate duplicated property names in some cases
 
 v2.15.6
 ---